+++ test/libsolidity/syntaxTests/natspec/docstring_non_public.sol
@@ -1,0 +1,7 @@
+contract C {
+    /// @dev example of dev
+    uint private state;
+    /// another example, that is implicit dev.
+    uint another;
+}
+// ----