+++ test/libsolidity/syntaxTests/natspec/docstring_private_state_variable.sol
@@ -1,7 +1,6 @@
 contract C {
     /// @notice example of notice
-    /// @dev example of dev
     uint private state;
 }
 // ----
-// DocstringParsingError 6546: (17-74): Documentation tag @notice not valid for non-public state variables.
+// DocstringParsingError 6546: (17-46): Documentation tag @notice not valid for non-public state variables.