+++ Changelog.md
@@ -13,6 +13,8 @@ Compiler Features:
  * SMTChecker: Report out of bounds index access for arrays and fixed bytes.
  * Standard JSON: Model checker option ``settings.modelChecker.targets`` also accepts ``outOfBounds``.
  * Yul Optimizer: Added a new step FunctionSpecializer, that specializes a function with its literal arguments.
+ * NatSpec: Allow NatSpec tag for non-public state variables. This documentation will only be part of the AST, under the field ``documentation``.
+ * NatSpec: The default tag for non-public state variable is now ``@dev``.
 
 
 Bugfixes: